chore: move class error
