+++ src/post-message.usecase.ts
@@ -1,5 +1,5 @@
 import {MessageRepository} from "./message.repository";
-import {Message} from "./message";
+import {Message, MessageCannotBeEmptyError, MessageCannotOnlyBeSpaceError, MessageTooLongError} from "./message";
 
 
 export type PostMessageCommand = Omit<Message, 'publishedAt'>;
@@ -9,26 +9,6 @@ export interface DateProvider {
     getNow(): Date;
 }
 
-export class MessageTooLongError extends Error {
-    constructor(message = 'Message too long') {
-        super(message);
-        this.name = 'MessageTooLongError';
-    }
-}
-
-export class MessageCannotOnlyBeSpaceError extends Error {
-    constructor(message = 'Message cannot only be space') {
-        super(message);
-        this.name = 'MessageCannotOnlyBeSpaceError';
-    }
-}
-
-export class MessageCannotBeEmptyError extends Error {
-    constructor(message = 'Message cannot be empty') {
-        super(message);
-        this.name = 'MessageCannotBeEmptyError';
-    }
-}
 
 export class PostMessageUseCase {
     constructor(
